Revert "Switch RSA_sign to size_t." (#684)
* Revert "Switch RSA_sign to size_t."

This reverts commit da59eeb9ea33c24abab7af35c1fb2b75d7462d3a.

* Revert harder

+++ crypto/fipsmodule/rsa/internal.h
@@ -148,7 +148,7 @@ int rsa_verify_raw_no_self_test(RSA *rsa, size_t *out_len, uint8_t *out,
                                 size_t in_len, int padding);
 
 int rsa_sign_no_self_test(int hash_nid, const uint8_t *digest,
-                          size_t digest_len, uint8_t *out, unsigned *out_len,
+                          unsigned digest_len, uint8_t *out, unsigned *out_len,
                           RSA *rsa);
